w3c-reports/12092014/gtk/runned-tests/2dcontext/text-styles/2d.text.draw.baseline.hanging-expected.txt

2d.text.draw.baseline.hanging
Actual output:
Expected output:


FAIL Canvas test: 2d.text.draw.baseline.hanging assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:47:25
http://localhost:8800/2dcontext/text-styles/2d.text.draw.baseline.hanging.html:37:23
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:843:37)
